visualization/requirements.txt
psycopg2-binary
folium
requests
flask
altair
pandas